Fix spacing in MANIFEST file
The manifest files in jars are apparently pretty picky with respect to
spacing and don't like having the spaces in front of the keys. If they
do the file is ignored and jar just generates a blank one, conveniently
leaving out the Main-Class directive making the jar useless. Given the
literal quotes in python the rest of the lines need to be moved over to
avoid adding extra whitespace.