Add properties to enable access to Lexer fields
In the Java runtime, Lexer actions can access fields such as _modeStack and _mode. These fields were marked public in Java, but are marked private in C#. I assume this is to preserve CLS compliance because the field names start with an underscore, therefore must be private. 

I needed to access these fields from the C# runtime in one of my grammars, so I hope you can implement this fix.

Additionally, I switched the _modeStack variable to implement an actual stack.