Report an end position when reporting syntax errors
Clients often want to report a span position for any errors. Currently,
an end position can be reconstructed from the token length for parser
errors, but lexer errors do not have tokens attached. Nonetheless, we do
know the ending index of the error, and it is helpful to report it.