fixes bug in visit()
In the common case where a rule has labeled productions, visit() would look for a function named visitX where X was the rule name, not the *label* name - and that function isn't even generated by the generator, since it generates functions for each labeled rule.

E.g. in the rule below, the visitor would be generated with functions visitWrap(), visitAdd(), and visitMult() - but the visit function logic would attempt to call visitExpr(), which is *not* generated.

expr
    : '(' expr ')'	# wrap
    | expr '+' expr	# add
    | expr '*' expr	# mult
    ;