If an instance of a class deriving from ParserRuleContext is created with an invokingState of 0 then the constructor will incorrectly change this to null due to the falsey nature of 0.
A similar problem exists in the base class of ParserRuleContext,
that is RuleContext, where it will change an invokingState of 0
to -1. Using the Java source as a reference, this appears to be a
bug. The invokingState should be set to -1 when the parent is null.

It looks like the code is trying to watch for an invokingState of
null or undefined. However, I think this is unnecessary since the
rule context classes generated from the grammar already handle
these cases, for example:

function ContentContext(parser, parent, invokingState) {
        if(parent===undefined) {
            parent = null;
        }
        if(invokingState===undefined || invokingState===null) {
                invokingState = -1;
        }