Replace init with init_ in DelegatorRuleOverridesLookaheadInDelegate.
init is the name of the instance initializer function in Swift, so generating
a grammar with that name causes problems, since it is generated into
function names on the parser.

I thought that renaming this would be a problem for other targets, but it
doesn't look like it, and actually the convention of using type_ rather than
type is already in use in this file.  Match this by adopting init_ as the
name, and regenerate the tests to match.

This was initially handled in https://github.com/janyou/ANTLR-Swift-Target
(using the name initIt instead of init).