Wrap some values with <OptionalValue> declarations.
I don't know what this is for.

I also presume that this requires corresponding changes in the other targets
that aren't there yet (per the comment at the bottom).  Just committing this
for now.

This comes from https://github.com/janyou/ANTLR-Swift-Target and is
marked Copyright (c) 2016 janyou on top of the BSD license and Copyrights
for Terence Parr and Sam Harwell derived from the original ANTLR source.