trim runtime-testsuite/pom.xml dependencies
TestGenerator's dependencies should be the only ones needed
in 'compile' scope.

Dependencies needed for compiling and executing the tests
can have 'test' scope.

The antlr Java runtime is in the test classpath, so does
not need to be copied as a test resource.

Removed master POM's testResources element.  It had caused
unnecessary copying of test source files.