fix TestGenerator.getOutputDir() NullPointerException in Windows
When running TestGenerator in Windows, it got a NullPointerException
in getOutputDir() because it was confused by a mixture of '/' and '\'
delimiters in the directory path.

This change lets TestGenerator accept any mixture of the two kinds
of slashes as directory delimiters when running in Windows, and
makes its pathname handling more portable and robust.