TestGenerator output on Windows should be same as other platforms.
Windows output was different from other platforms because
Java's default line.separator is different: \r\n on Windows,
\n everywhere else.  This is an unnecessary anachronism.
Since the blessed departure of MSDOS and Windows 98, almost
all Windows software has accepted portable \n line endings
interchangeably with the antiquarian \r\n curiosities.

Some test templates generate code like

  StringBuilder buf = new StringBuilder(61);
  buf.append(string_of_exactly_61_chars);

On Windows, the buffer size would come out as 64 instead of
61 if the string contains 3 newlines.  Each useless \r is
stripped from the literal but this happens after sizing the
buffer.  Thus TestGenerator produced meaningless changes in
the generated code depending on which platform was used.
The buffer size isn't critical, but when anything changes,
it takes someone's time to see why it changed and be sure
the change is ok.

The solution is to specify \n as the line.separator for all
platforms.  It has to be specified when launching the JVM,
since in newer Java releases it is read-only.