TestGenerator options to include/exclude selected targets
TestGenerator now accepts multiple occurrences of the
"-target <name>" option and the new "-exclude <regex>"
option.  This should make it more convenient to generate
exactly the desired set of targets.  Formerly the -target
option could be used only once.

The "-exclude <regex>" option skips generating tests for
target names that fully match the pattern.  The matching
is case insensitive.

As a shorthand, these options accept multiple args. Each
arg but the last ends with a comma: "-target <name>[, ...]"
or "-exclude <regex>[, ...]".  "-target Java, Python3" is
the same as "-target Java -target Python3".

The behavior described below is unchanged, except as noted.

Target names are case-insensitive on Windows, but on other
platforms they can be case-sensitive in certain contexts -
so for portability, developers on Windows should observe
ANTLR's spelling of target names, with initial capitals.

"-target ALL" (case-insensitive) expands to TestGenerator's
built-in list of targets, not including the browsers.  ALL
is the default.

The -target option can specify a <name> that is not in the
built-in list.  TestGenerator will look for a file named
resources/org/antlr/v4/test/runtime/<name>/<name>.test.stg
where <name> is lower cased in the first occurrence and
mixed case (as given) in the second. (Case-insensitive on
Windows only!)  If <name> has a slash - <a>/<b> - the file
is resources/org/antlr/v4/test/runtime/<a>/<b>/<b>.test.stg

If the "-browsers" option is given, TestGenerator's
built-in list of browser targets is added to the target set.
"-browsers" formerly was ignored unless "-target ALL" was
specified or assumed, but with this change, the two options
are independent and you must specify both if you want the
complete set.

The order of these options doesn't matter.  -exclude is
processed last.