tool-testsuite test-jar cleanup
It seems the tool tests were formerly built by tool/pom.xml
and placed in a test-jar, from where they were executed by
tool-testsuite/pom.xml.  Later it was changed so that the
tests are both compiled and executed in tool-testsuite.  The
test-jar was still created but was empty.

With this patch the empty test-jar is no longer created and
the dependency on it is removed.

A (nonempty) test-jar now is built by tool-testsuite/pom.xml
in case it is wanted for some reason.  (It isn't needed for
build/test.)