tidy up pom.xml <resources> and <testResources>
Resources which are needed only for testing are now
tagged as <testResources> rather than <resources>.

The tool-testsuite and runtime-testsuite subprojects
consist of tests only, with no runtime artifacts, so
they use <testSourceDirectory> and <testResources>
but not <sourceDirectory> nor <resources>.

Not all subprojects use the same layout for source,
resource, and test directories - so remove them from
the master pom.xml  and let each subproject specify
them where needed.

In tool-testsuite/pom.xml, added an <include> so that
a lot of files aren't copied unnecessarily.

The runtime-testsuite/pom.xml did and still does copy
a lot of files from the runtime subproject that aren't
actually needed or used.  Added an <exclude> to abate
unnecessary copying for the Java target.  The other
targets remain as opportunities for improvement.