Java: shrink critical section in LexerATNSimulator
When using the same Lexer in many threads, the LexerATNSimulator.addDFAState
can be a bottleneck. The most expensive work in this section is computing the
hashCode of the ATNConfigSet. This happens twice, once when checking if the
state is already in the map, and once when inserting it. If we make the configs
readonly before both operations, the hashCode will be reused. Second, if we
compute the hashCode outside of the critical section it will be cached,
shortening the total time in that section.