[C++] Fix building shared library in mingw toolchain
This fixes building shared library by defining proper symbols.
Previously, the symbols ANTLR4CPP_EXPORTS and ANTLR4CPP_STATIC
were defined for MSVC toolchain only. So, mingw toolchain was
excluded in that condition. Hence, the issue is solved by
moving the compile definitions in a general `if(WIN32)` check
which is independent of compiler toolchain.

Signed-off-by: Biswapriyo Nath <nathbappai@gmail.com>