Issue #3988: Builds failing on Ubuntu
The workflow is setup to work on Ubuntu-20.04 and the default 'latest'
is migrating to Ubuntu-22.04 which isn't supported yet (all the
dependent actions needs to support the platform before the workflow can
use it).

Switching all platforms to use explicit versions rather than the
'-latest' tag so the transition is more voluntary than something that
just happens behind the scene.

Also, upgraded all setup-xx actions to latest versions to remove some of
the deprecation warnings on dashboard.

Setup concurrency as well (kill any existing build running on the same
branch branch and using the same workflow).

Signed-off-by: HS <hs@apotell.com>