Python2+3 runtimes: Define getTokenNames.
Fixes #1043: a list of token names can be easily formulated by combining
the lists of literalNames and symbolicNames, since they use the same
indexing system.

Also similarly fixes the absence of getRuleNames.

Also fixes attempting to use the result of zip() as a map which broke
both getTokenTypeMap and getRuleIndexMap. In Python 2, zip() returns
a list of tuples. In Python3, zip returns an object that can only be
iterated over once (and thus would not be useful to cache).

Signed-off-by: Benjamin S Wolf <bswolf@google.com>