Cpp: Remove code duplication (#3995)
* Cpp: Remove code duplication

The same function was defined inside multiple different source files (in
an anonymous namespace).
Not only is this generally bad practice from a maintenance point of
view, but it also break unity builds (which can speed up compilation
times A LOT).

The fix simply consists in introducing a new Utils file that contains
the single implementation and every source file that uses this function
simply has to be linked against that new file (and include the
corresponding header).

Signed-off-by: Robert Adam <dev@robert-adam.de>

* CI: Enable unity builds for cpp targets

This should reduce the build time for those steps significantly

Note: Unity builds are enabled only for cmake-based CI builds

Signed-off-by: Robert Adam <dev@robert-adam.de>

* CI: Perform unity and non-unity cpp builds

Signed-off-by: Robert Adam <dev@robert-adam.de>

---------

Signed-off-by: Robert Adam <dev@robert-adam.de>
Signed-off-by: Jim.Idle <jimi@idle.ws>