Issue #4185: Too many artifacts fail to upload
Github action for upload was upgraded to v3 recently and the release is
unstable causing too many uploads to fail. Downgrading back to previous
version hasn't made significant improvement either.

Since the artifacts aren't exactly used by any chained job, failures for
uploading the artifact can be ignored. The artifacts are used mostly for
the purpose for debugging and so if needed the user can trigger specific
build again to get the artifact.

Signed-off-by: HS <hs@apotell.com>