feat: Stop using interfaces for ATNConfigSets
Too much emulation of the Java runtime structure means that many structs
are being called via interfaces, which introduces and extra look up. When
you add that lookup to call via a pointer and then even an extra lookup for
generics, then it becomes significant. This seems like a big commit but
it is all just changing the one type declaration.

This now allows us to get rid of the stupid getters and setters that just
clutter the code and are not idiomatic of Go anyway.

Signed-off-by: Jim.Idle <jimi@idle.ws>