fix: Speed up  test suites
  - Upgrades surefire
  - Limits amount of parallelism to stop swamping the test machines.

Because the test suite does not strictly follow the jupiter config, which was
configured to use as many cores as possible, the machine would end up swamped
and either IO bound, or just defeating any CPU cache advntages.

The change to limit parallelism to 3 means that everythign goes faster,
including the Java tests, which do not need to create external processes.

This took some working out.

Signed-off-by: Jim.Idle <jimi@idle.ws>