Generate meaningful exception message on checkVersion() failure
Currently, loading a parser from an incompatible version results in a message
like:

  Could not deserialize ATN with version  (expected 4).

(see e.g. https://github.com/antlr/antlr4/issues/4041 )

Change the exception message generation to use ord(version), so it shows e.g.
the string 3 instead of the character with ordinal 3, which is non-printable
and the cause for the above cryptic message.

Signed-off-by: Raffaello D. Di Napoli <rafdev@dinapo.li>