[TS] Add property versions of raw fields to TS
JS properties exist for the `_type` and `_input` fields, but they aren't
present in the typescript definition file. This commit adds those properties.

The commit adds doc comments pointing developers toward the corresponding
properties. I don't know if there's a deprication process for removing the
old fields. `inputStream`'s setter function does a reset of the lexer, but
otherwise the properties are normal.

Signed-off-by: Phlosioneer <mattmdrr2@gmail.com>