Fix "chrono" C++ library not being included on Windows
On Linux, the build is successful because this library is usually included in the headers by default. But on Windows, when using Visual Studio (at least VS 2022 Community - the version I tested this on), the build fails because the "chrono" library is not included by default.

To avoid troubling new Windows adopters of this C++ runtime, it would be good to include this in the code!

(Adding this line does not break builds on Linux, and does not cause any warnings on Windows or Linux! - I test it this on both Windows and Linux)

Signed-off-by: 0xENDER <72353925+0xENDER@users.noreply.github.com>